# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for the EatMe recipe generation service",
        "title": "EatMe API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/recipes": {
            "get": {
                "description": "Get all recipes for user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipe"
                ],
                "summary": "Get all recipes for user",
                "operationId": "getAllRecipes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserRecipe"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/recipes/{recipe_id}": {
            "get": {
                "description": "Get recipe by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipe"
                ],
                "summary": "Get recipe by ID",
                "operationId": "getRecipe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recipe ID",
                        "name": "recipe_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserRecipe"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "404": {
                        "description": "Recipe not found",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete recipe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipe"
                ],
                "summary": "Delete recipe",
                "operationId": "deleteRecipe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recipe ID",
                        "name": "recipe_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserRecipe"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "404": {
                        "description": "Recipe not found",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.APIError": {
            "description": "APIError represents an API error response",
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.Ingredient": {
            "description": "Ingredient represents an ingredient in a recipe",
            "type": "object",
            "required": [
                "name",
                "quantity",
                "unit"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Flour"
                },
                "quantity": {
                    "type": "number",
                    "example": 1
                },
                "unit": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.MeasurementUnit"
                        }
                    ],
                    "example": "cup"
                }
            }
        },
        "models.MeasurementUnit": {
            "type": "string",
            "enum": [
                "g",
                "ml",
                "tsp",
                "tbsp",
                "cup",
                "oz",
                "lb"
            ],
            "x-enum-varnames": [
                "MeasurementUnitGram",
                "MeasurementUnitMilliliter",
                "MeasurementUnitTeaspoon",
                "MeasurementUnitTablespoon",
                "MeasurementUnitCup",
                "MeasurementUnitOunce",
                "MeasurementUnitPound"
            ]
        },
...
        "models.UserRecipe": {
            "description": "UserRecipe is the user's personal copy (favorites, edits).",
            "type": "object",
            "required": [
                "created_at",
                "description",
                "id",
                "ingredients",
                "is_favorite",
                "latest_version_id",
                "servings",
                "steps",
                "thread_id",
                "title",
                "total_time_minutes",
                "updated_at",
                "user_id"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "example": "A classic Italian dish"
                },
                "global_recipe_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Ingredient"
                    }
                },
                "is_favorite": {
                    "type": "boolean"
                },
                "latest_version_id": {
                    "type": "string"
                },
                "servings": {
                    "type": "integer",
                    "example": 4
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "thread_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "example": "Veal Bolognese"
                },
                "total_time_minutes": {
                    "type": "integer",
                    "example": 120
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}